[kyo-system] move Path, System, Process, and Command out of kyo-core
Relocation only: no public signature changes and no downstream source edits.
The stream file sinks move alongside them because they are typed in terms of
Path, and leaving them in kyo-core would make kyo-core depend on kyo-system.
