+++ kyo-system/js-wasm/src/main/scala/kyo/internal/SystemPlatformSpecific.scala
