+++ kyo-system/jvm-native/src/main/scala/kyo/internal/PathPlatformSpecific.scala
