+++ kyo-system/jvm-native/src/main/scala/kyo/internal/SystemPlatformSpecific.scala
