+++ kyo-system/shared/src/main/scala/kyo/CommandException.scala
