+++ kyo-system/shared/src/main/scala/kyo/FileException.scala
