+++ kyo-system/shared/src/test/scala/kyo/PathSizeTest.scala
