+++ kyo-system/shared/src/test/scala/kyo/PathStatTest.scala
